requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==3.7.1
asyncer==0.0.2
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
bangla==0.0.2
bidict==0.23.1
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chainlit==1.1.400
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cloudpathlib==0.18.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses-json==0.5.14
dateparser==1.1.8
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
fastapi==0.110.3
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
grpcio==1.65.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
inflect==7.3.1
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.24
langchain-openai==0.1.19
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.93
language_data==1.2.0
Lazify==0.4.0
lazy_loader==0.4
librosa==0.10.2.post1
literalai==0.0.607
llama_cpp_python==0.2.84
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
onnx==1.16.2
onnx-simplifier==0.4.36
onnxruntime==1.18.1
openai==1.37.1
opencv-python==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.3
packaging==23.2
pandas==1.5.3
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
protobuf==4.25.4
psutil==6.0.0
pushnotifier==1.3.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
pynndescent==0.5.13
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
regex==2024.7.24
requests==2.32.2
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.1
syncer==2.0.3
tenacity==8.3.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
trainer==0.0.36
transformers==4.43.3
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzlocal==5.2
ultralytics==8.2.73
ultralytics-thop==2.0.0
umap-learn==0.5.6
Unidecode==1.3.8
uptrace==1.26.0
urllib3==2.2.1
uvicorn==0.25.0
wasabi==1.1.3
watchfiles==0.20.0
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
